Takes consideration of patient centric images like DICOM format.
In the case of DICOM images, axes are not related to a "world' but to the patient:
X direction is from the right hand to the left hand of the patient.
Y direction is from Anterior to Posterior.
Z direction is from Feet to Head.
For a proper management of origin in image actors, we need to take the Image Orientation tag in consideration.
Behavior of others classical images (Analyse, Interfile..) is not altered by this commit.